+++ ExtraAddIns/UserStream/UserStreamAddIn.cs
@@ -289,7 +289,7 @@ namespace Misuzilla.Applications.TwitterIrcGateway.AddIns.UserStream
         {
             return new User()
                        {
-                           Id = (Int32)this.id,
+                           Id = (Int64)this.id,
                            Protected = this.Protected,
                            ProfileImageUrl = this.profile_image_url,
                            ScreenName = this.screen_name
